refactor: Move strs to public linter package
