+++ internal/addon/rules/fileNamesLowerSnakeCaseRule.go
@@ -7,8 +7,8 @@ import (
 	"github.com/yoheimuta/protolint/internal/stringsutil"
 
 	"github.com/yoheimuta/go-protoparser/parser"
-	"github.com/yoheimuta/protolint/internal/strs"
 	"github.com/yoheimuta/protolint/linter/report"
+	"github.com/yoheimuta/protolint/linter/strs"
 	"github.com/yoheimuta/protolint/linter/visitor"
 )
