+++ linter/strs/strs.go
